requirements.txt
absl-py==2.1.0
accelerate==1.4.0
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
asyncpg==0.30.0
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.13.1
black==25.1.0
bleach==6.1.0
blinker==1.9.0
boolean.py==5.0
bs4==0.0.2
build==1.1.1
cachetools==5.3.3
certifi==2025.1.31
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
ci-info==0.3.0
click==8.1.7
clingo==5.8.0
cloudpickle==3.0.0
cohere==5.2.2
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.8
configparser==7.1.0
contourpy==1.2.0
coverage==7.7.1
cryptography==45.0.5
cvxopt==1.3.2
cycler==0.12.1
dash==3.0.1
dataclasses-json==0.6.4
datasets==3.3.1
dateparser==1.2.1
debugpy==1.8.1
declare4py==2.2.0
decorator==5.1.1
deepeval==2.6.5
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dice-ml==0.11
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docopt==0.6.2
docstring_parser==0.16
docutils==0.21.2
ecdsa==0.19.1
elastic-transport==8.12.0
elasticsearch==8.12.1
elasticsearch-dsl==8.12.0
et-xmlfile==1.1.0
etelemetry==0.3.1
execnet==2.1.1
executing==2.0.1
fastapi==0.116.1
fastavro==1.9.4
fastjsonschema==2.19.1
filelock==3.13.3
filetype==1.2.0
fire==0.6.0
fitz==0.0.1.dev2
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.50.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
funcy==2.0
future==1.0.0
gast==0.5.4
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.29.0
google-pasta==0.2.0
googleapis-common-protos==1.69.2
graphviz==0.20.2
greenlet==3.1.1
grpcio==1.71.0
h11==0.16.0
h5py==3.12.1
hf-xet==1.1.5
holidays==0.72
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
hugchat==0.4.5
huggingface-hub==0.33.4
humanfriendly==10.0
hyperopt==0.2.7
idna==3.6
imageio==2.36.0
importlib-metadata==6.11.0
importlib_resources==6.4.0
iniconfig==2.1.0
inquirerpy==0.3.4
install==1.3.5
instructor==1.7.7
intervaltree==3.1.0
ipykernel==6.29.3
ipython==8.12.3
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.8.2
joblib==1.3.2
json5==0.9.22
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.1
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.8.0
keyring==25.6.0
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.3.21
langchain-cohere==0.1.0
langchain-community==0.3.20
langchain-core==0.3.48
langchain-elasticsearch==0.1.1
langchain-openai==0.3.10
langchain-text-splitters==0.3.7
langsmith==0.3.18
lark==1.2.2
lark-parser==0.12.0
lazy_loader==0.4
libclang==18.1.1
lime==0.2.0.1
llama-cloud==0.1.16
llama-cloud-services==0.6.8
llama-index==0.12.25
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.25
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.9
llama-index-llms-openai==0.3.27
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.4.post1
llama_stack==0.2.15
llama_stack_client==0.2.15
llvmlite==0.42.0
logaut==0.2.0
looseversion==1.3.0
ltlf2dfa==1.0.2
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
MechanicalSoup==1.3.0
mistune==3.0.2
ml_dtypes==0.5.1
mlxtend==0.23.4
mmh3==4.1.0
monotonic==1.6
more-itertools==10.6.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.7
narwhals==1.27.1
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.2
nest-asyncio==1.6.0
networkx==3.2.1
nh3==0.2.21
nibabel==5.2.1
nipype==1.8.6
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.7
onnxruntime==1.17.1
openai==1.95.1
openai-whisper==20240930
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opentelemetry-util-http==0.44b0
opt-einsum==3.3.0
optree==0.10.0
orjson==3.9.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.3
pathlib==1.0.1
pathspec==0.12.1
patsy==0.5.6
pdf2docx==0.5.8
pexpect==4.9.0
pfzy==0.3.4
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
pkginfo==1.10.0
platformdirs==4.2.0
plotly==6.0.0
pluggy==1.5.0
pm4py==2.7.12.4
portalocker==3.1.1
posthog==3.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==5.29.4
prov==2.0.1
psutil==5.9.8
psycopg2==2.9.10
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
py4j==0.10.9.7
pyaml==25.7.0
pyarrow==19.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.1
pydotplus==2.0.2
PyGetWindow==0.0.9
Pygments==2.17.2
pylogics==0.2.1
PyMsgBox==1.0.9
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyobjc-core==10.2
pyobjc-framework-Cocoa==10.2
pyobjc-framework-Quartz==10.2
pyogrio==0.9.0
pyparsing==3.1.2
pypdf==5.4.0
PyPDF2==3.0.1
pyperclip==1.8.2
PyPika==0.48.9
pyproj==3.6.1
pyproject_hooks==1.0.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-repeat==0.9.3
pytest-rerunfailures==12.0
pytest-xdist==3.6.1
pythomata==0.3.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.5.0
python-json-logger==2.0.7
python-multipart==0.0.20
pytweening==1.2.0
pytz==2025.2
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==25.1.2
raiutils==0.4.2
rdflib==6.3.2
readme_renderer==44.0
referencing==0.33.0
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
rubicon-objc==0.4.7
safetensors==0.5.2
schedule==1.2.2
scikit-image==0.24.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
selenium==4.18.1
Send2Trash==1.8.2
sentence-transformers==5.0.0
sentencepiece==0.2.0
sentry-sdk==2.24.1
setuptools==69.2.0
shap==0.46.0
shapely==2.0.5
simplejson==3.19.3
six==1.16.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
sns==0.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.47.1
statsmodels==0.14.4
streamlit==1.42.1
striprtf==0.0.26
sympy==1.12
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.4.0
terminado==0.18.1
tf_keras==2.19.0
threadpoolctl==3.3.0
tifffile==2024.9.20
tiktoken==0.9.0
tinycss2==1.2.1
tokenizers==0.21.0
toml==0.10.2
torch==2.2.2
tornado==6.4
tqdm==4.67.1
traitlets==5.14.2
traits==6.3.2
transformers==4.49.0
trio==0.25.0
trio-websocket==0.11.1
twine==5.1.1
typer==0.10.0
types-python-dateutil==2.9.0.20240315
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.35.0
uvloop==0.19.0
virtualenv-clone==0.5.7
watchdog==6.0.0
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.0.3
xlrd==2.0.1
xxhash==3.5.0
yarg==0.1.9
yarl==1.9.4
zipp==3.18.1
zstandard==0.23.0
